CSS theme support
The plugin now has some support for CSS to modify the appearance of the dock. You can set the indicator style to 'None' to use only CSS rather than any of the built in indicator styles. Using custom CSS in combination with any of the indicators is supported but may cause some issues.

See `themes/example.css` for a really scuffed example of how to use it.