Speed up ftw() by increasing handle limit
What are the effects of traversing the directory tree in
findXDGDirectories() (at startup time) rather than loadXDGDirectories()
(also called on .desktop rescan)? It might be faster, since we only do a
recursive rescan once, but it's flawed since we don't pick up changes to
*which* directories are present.

One theoretical advantage of having a recursive directory scan, followed
by a linear file scan, is that if XDG_DATA_DIRS contains both a
directory and its child, we won't process those .desktop files twice.
But in practice we can avoid duplicated files (and also be more
standards-conformant [1]) by not pushing /usr/local/share and /usr/share
if we find a non-empty XDG_DATA_DIRS.

[1]: https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html#variables