Fix reordering (ctrl+dragging) icons to the right
Previously dragging a launcher towards the right would leave it 1 icon
to the left of the position you dragged it to. And you couldn't drag a
launcher to the rightmost position in the dock. This fixes it.

I'm not sure why the -1 was added in the first place. Is there now a
risk of now dropping an icon at position "length", which is an invalid
position? I haven't seen any malfunctions in my testing though.