Store pinned apps as ids instead of paths
There doesn't seem to be any good reason to use the path rather than the
id, which is the real defining element of an app. Especially since
b11cc9cc that fixes desktop file override. This also gets rid of the
path update code when editing the launcher: the id remains the same and
the other attributes are updated by monitoring the app directories.