AppInfos: Use std::shared_ptr to manage memory allocation
The raw pointer is now only used in callbacks; every object using an
AppInfo takes a reference to it. With the changes made in previous
commits, this is enough to fix memory leaks in AppInfos.cpp. But there
are still objects referencing AppInfo that are not released. This is the
subject of next commits.

Helps: #19