Release data as much as possible on XfcePanelPlugin::free-data signal
This is required when the plugin is run as internal, otherwise certain
wnck signals remain connected and cause the panel to crash after the
plugin is removed. Beyond that, memory should be freed as much as
possible, which should be the case now even if the somewhat abusive use
of namespaces doesn't allow this to be done thoroughly.