Helpers: Always stop timeout before starting a new one
As currently implemented, the timeout doesn't support parallel launches,
and is unlikely to be needed in the future. In particular, this fixes a
bug that caused the window menu to disappear in certain cases (two menu
hiding timeouts launched successively, and only one is stopped when the
menu is shown).