Fix blurry icons/thumbnails when UI scale > 1
This also fixes e5f8d337 where the dnd icon was not scaled.

Unfortunately, the widget and icon size updates don't fully allow you
to change the scale factor at runtime. It works well from factor 1 to 2
or more, but in the opposite direction the window menu remains badly
positioned. This is due to a bug in GtkSocket/GtkPlug: the geometry of
the GdkWindow returned by gtk_plug_get_socket_window() is not updated
after scaling, so xfce_panel_plugin_position_widget() returns an
erroneous position. Obviously, this is not an issue if the plugin is run
as internal.