x11: Add exceptions where instance-id should be preferred to class-id
It seems safe to add these exceptions, because if you have this kind of
generic name as class-id, the instance-id should probably provide
something more specific, or simply confirm the class-id. And if the app
changes class-id in the future to a valid/specific one, this match will
fail and the new class-id will be well used.

Fixes: #75